DARC_CLIENT/requirements.txt
pyqt6
pycryptodome
websockets
qiskit
qiskit-aer
numpy
